pyproject.toml
[tool.poetry]
name = "medieval-deck"
version = "0.1.0"
description = "A roguelike card game with medieval theme"
authors = ["Bruno"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygame = "^2.5"
diffusers = "^0.25"
torch = "^2.1"
pillow = "^10.0"
numpy = "^1.26"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-pygame = "^2.0"
ruff = "^0.4"
black = "^24.4"
isort = "^5.13"
mypy = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
